backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
networkx
groq
pydantic
python-dotenv
